fix(inventory): stop fractional stock truncating in dashboard + Zobaze import
Self-review sweep after the fractional-stock migration caught two spots that
still forced integers:

- Dashboard overview read SUM(stock) as int (InventoryMetrics.totalStock) —
  now that stock is REAL this could throw or truncate. Made totalStock double,
  read<double>, and formatted the display.
- Zobaze import parsed imported STOCK with _int (truncating e.g. 1.5 -> 1);
  now uses _num so fractional imported stock is preserved.

Verified: analyze clean, 141 tests pass, release APK 83.9MB.

+++ apps/mobile_flutter/lib/core/database/mobile_repository.dart
@@ -439,7 +439,7 @@ class InventoryRepository {
         .map((row) {
           final metrics = InventoryMetrics(
             totalItems: row.read<int>('total_items'),
-            totalStock: row.read<int>('total_stock'),
+            totalStock: row.read<double>('total_stock'),
             inventoryValue: row.read<double>('inventory_value'),
             potentialProfit: row.read<double>('potential_profit'),
             lowStock: row.read<int>('low_stock'),
